backend/requirements.txt
# Khaznati DZ Backend Dependencies

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2

# Session Management
starlette-session==0.4.3

# Storage (S3)
boto3==1.34.34
aioboto3==12.3.0

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP Client (for Supabase API)
httpx==0.26.0
aiohttp==3.9.3

# Email
aiosmtplib==3.0.1

# Utilities
python-dotenv==1.0.1
orjson==3.9.13

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9
